- renaming package to be called 'ops' from 'LL' - Using Generic for BinaryTree.Node.data instead hard-coding it to be of type int

+++ src/com/selfevaluation/ops/queue/QueueOps.java
@@ -127,7 +127,7 @@ public class QueueOps {
 
             while (n>0)
             {
-                data = dequeue().getData();
+                data = dequeue().getData().intValue();
                 generatedNumbers[originalLength-n] = data+"";
 
 //                enqueue(new Node(node.getData()+"0"));
